@@ -12,8 +12,7 @@ exports.config = {
     'browserName': 'chrome'
   },
   directConnect: true,
-   baseUrl: 'http://localhost:4200/',
-  //baseUrl: 'http://iorder.calibrecue.com/',
+  baseUrl: 'http://localhost:4200/',
   framework: 'jasmine',
   jasmineNodeOpts: {
     showColors: true,
@@ -22,7 +21,7 @@ exports.config = {
   },
   onPrepare() {
     require('ts-node').register({
-      project: require('path').join(__dirname, './tsconfig.e2e.json')
+      project: require('path').join(__dirname, './tsconfig.json')
     });
     jasmine.getEnv().addReporter(new SpecReporter({ spec: { displayStacktrace: true } }));
   }
