@@ -1,21 +1,16 @@
 import { NgModule } from '@angular/core';
-import {enableProdMode} from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { CustomerComponent } from './pages/customer/customer.component';
-import { LoaderComponent } from './pages/loader/loader.component';
+import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 
-enableProdMode();
 const routes: Routes = [
-  { path: '', redirectTo: 'customer', pathMatch: 'full' },
-   { path: 'customer/:tableName/:CompanyID', component: CustomerComponent },
-  // { path: '/customer/:tableName/:CompanyID/', component: CustomerComponent },
-  { path: 'loader', component: LoaderComponent },
+  {
+    path: '',
+    loadChildren: () => import('./tabs/tabs.module').then(m => m.TabsPageModule)
+  }
 ];
-
 @NgModule({
- 
-  imports: [ RouterModule.forRoot(routes) ],
-  exports: [ RouterModule ]
+  imports: [
+    RouterModule.forRoot(routes, { preloadingStrategy: PreloadAllModules })
+  ],
+  exports: [RouterModule]
 })
 export class AppRoutingModule {}
-
